add take order support to UI components
- Add TAKE_ORDER transaction type
- Add createTakeOrderTransaction to TransactionManager
- Add Take Order button to OrderDetail component
- Add Take button and Actions column to OrdersListTable
- Add onTakeOrder callback prop to OrderDetail
- Add handleTakeOrderModal prop to OrdersListTable

+++ packages/ui-components/src/lib/components/detail/OrderDetail.svelte
@@ -67,6 +67,13 @@
 		| ((raindexClient: RaindexClient, vaultsList: RaindexVaultsList) => void)
 		| undefined = undefined;
 
+	/** Callback function when take order action is triggered
+	 * @param order The order to take
+	 */
+	export let onTakeOrder:
+		| ((raindexClient: RaindexClient, order: RaindexOrder) => void)
+		| undefined = undefined;
+
 	let codeMirrorDisabled = true;
 	let codeMirrorStyles = {};
 
@@ -125,14 +132,19 @@
 			</div>
 
 			<div class="flex items-center gap-2">
-				{#if matchesAccount(data.owner)}
-					{#if data.active}
-						<Button
-							on:click={() => onRemove(raindexClient, data)}
-							data-testid="remove-button"
-							aria-label="Remove order">Remove</Button
-						>
-					{/if}
+				{#if matchesAccount(data.owner) && data.active}
+					<Button
+						on:click={() => onRemove(raindexClient, data)}
+						data-testid="remove-button"
+						aria-label="Remove order">Remove</Button
+					>
+				{/if}
+				{#if data.active && onTakeOrder}
+					<Button
+						on:click={() => onTakeOrder(raindexClient, data)}
+						data-testid="take-order-button"
+						aria-label="Take order">Take Order</Button
+					>
 				{/if}
 
 				<Refresh
